Change collisions
-Создан отдельный класс, обрабатывающий коллизии при столкновении с мячом.
-Удалены или исправлены разрозненные скрипты, в которых содержались коллизии.
