+++ Plate and ball/Assets/Scenes/Level1.unity
@@ -1930,7 +1930,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1734008643}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8a15d0f9aa380e24ebfa40cdfe67df20, type: 3}
+  m_Script: {fileID: 11500000, guid: 0c2063cfd0286584084e01e31a5b2786, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &1746541306
