+++ Plate and ball/Assets/Scripts/Ball/MoveBall.cs
@@ -6,9 +6,8 @@ public class MoveBall : MonoBehaviour, IMoveObject
 {
     [SerializeField] private Transform parentObject;
 
-    private const float force = 300.0f;
+    private const float FORCE = 300.0f;
 
-    private float lastPositionBall;
     private Rigidbody2D rb;
     private bool isActive;
     private ObjectInputController inputStartMove;
@@ -27,33 +26,18 @@ public class MoveBall : MonoBehaviour, IMoveObject
         {
             Move();
         }
-
     }
     public void Move()
     {
-        lastPositionBall = transform.position.x;
         isActive = true;
         transform.SetParent(null);
         rb.bodyType = RigidbodyType2D.Dynamic;
-        rb.AddForce(Vector2.up * force);
+        AddForce();
     }
 
-    public void OnCollisionEnter2D(Collision2D collision)
+    public void AddForce(float direction = 0, float differenceOfPositions = 0)
     {
-        float newPositionBall = transform.position.x;
-
-        if (collision.gameObject.TryGetComponent(out MovePlate ball))
-        {
-            if (newPositionBall > lastPositionBall - 0.1 && lastPositionBall + 0.1 > newPositionBall)
-            {
-                rb.velocity = Vector2.zero;
-                float playerCenterPosition = parentObject.GetComponent<Transform>().position.x;
-                float differenceOfPositions = playerCenterPosition - newPositionBall;
-                float direction = newPositionBall < playerCenterPosition ? -1 : 1;
-                rb.AddForce(new Vector2(direction * Mathf.Abs(differenceOfPositions * force), force));
-            }
-        }
-
-        lastPositionBall = newPositionBall;
+        rb.velocity = Vector2.zero;
+        rb.AddForce(new Vector2(direction * Mathf.Abs(differenceOfPositions * FORCE), FORCE));
     }
 }
