--- Plate and ball/Assets/Scripts/Block/BlockCollision.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class BlockCollision : MonoBehaviour
-{
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        if (collision.gameObject.TryGetComponent(out Block block))
-        {
-            block.ApplyDamage();
-        }
-    }
-}
