+++ Plate and ball/Assets/Scripts/UIScore.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8a15d0f9aa380e24ebfa40cdfe67df20
+guid: faa04ff84fe056248b81a4a257d233fc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
